#664 :recycle: refactor api calls to composable
